task 101

+++ 0x0C-more_malloc_free/101-mul.c
@@ -30,7 +30,7 @@ int main(int argc, char *argv[])
 			}
 		}
 	}
-	nul = atoi(argv[1]) * atoi(argv[2]);
+	nul = atol(argv[1]) * atol(argv[2]);
 	printf("%lu\n", nul);
 	return (0);
 }
